+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
 ## [6.4.0-rc.1] - 2026-09-02
 
 ### Added
@@ -850,6 +852,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[Unreleased]: https://https://github.com/internetguru/laravel-model-browser/compare/staging...dev
 [6.4.0-rc.1]: https://github.com/internetguru/laravel-model-browser/releases/tag/v6.3.0
 [6.3.0]: https://https://github.com/internetguru/laravel-model-browser/compare/v6.2.0...v6.3.0
 [6.3.0-rc.1]: https://github.com/internetguru/laravel-model-browser/releases/tag/v6.2.0